Move TestBindingDelegate to a testutil location
Also removing The `Experimental` annotation from `HandleManager` since we
are using it for real.
