+++ javatests/arcs/android/storage/handle/BUILD
@@ -36,11 +36,11 @@ arcs_kt_android_test_suite(
         "//java/arcs/core/storage/driver",
         "//java/arcs/core/storage/handle",
         "//java/arcs/core/storage/referencemode",
-        "//java/arcs/core/testutil",
         "//java/arcs/core/type",
         "//java/arcs/core/util",
         "//java/arcs/sdk/android/storage",
         "//java/arcs/sdk/android/storage/service",
+        "//java/arcs/sdk/android/storage/service/testutil",
         "//third_party/android/androidx_test/core",
         "//third_party/android/androidx_test/ext/junit",
         "//third_party/java/androidx/appcompat",
